Required changes
